+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Consume with error handling
+- Fix to work without a tracing
 
 ## 16.2.0 - 2021-09-14
 - Update dependencies